update: useVaccinationQuery -> useVacacineRecordQuery로 파일이름 변경 및 vaccineRecordQuery 추가
